#2447 Refactored identifier check so invalid characters are replaced later in the code generation process
